% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alternative_syns_vector.R
\name{alternative_syns_vector}
\alias{alternative_syns_vector}
\title{Produces alternative species names vector}
\usage{
alternative_syns_vector(search_syn_in_ncbi_out)
}
\arguments{
\item{search_syn_in_ncbi_out}{takes the output from \code{search_syn_in_ncbi}}
}
\value{
data.frame cols: facultative synonnmys, obligate synonyms and a combined new vector
}
\description{
Take the output from search_syn_in_ncbi and produces and alternative species vector
substituting the species not present on NCBI GenBank by synonyms that are present
}
\details{
This function is motivated as a helper for producing an input for MEGAPTERA. MEGAPTERA takes an input taxa dataset and automatically procecces downloaded sequences. However, it does not look for synonyms, which is a problem - especially for fungi. alternative_syns_vector, as the name says takes synonyms as input and produces a new vector of species names but subsitures "blind" species, thus species not present on GenBank but the synonyms are.
}
\examples{
taxa <- c("Heterobasidion annosum","Schizopora paradoxa")
s <- search_syn_in_ncbi(taxa)
spec <- alternative_syns_vector(search_syn_in_ncbi_out = s)
spec$syns
}
\author{
Franz-Sebastian Krah
}

