\name{rmycobank-package}
\alias{rmycobank-package}
\alias{rmycobank}
\docType{package}
\title{
\packageTitle{rmycobank}
}
\description{
\packageDescription{rmycobank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rmycobank}
\packageIndices{rmycobank}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rmycobank}

Maintainer: \packageMaintainer{rmycobank}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
