% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_syn_in_ncbi.R
\name{search_syn_in_ncbi}
\alias{search_syn_in_ncbi}
\title{Searches for the synonym on GenBank}
\usage{
search_syn_in_ncbi(taxon)
}
\arguments{
\item{taxon}{a \code{character} containing fungal species name}
}
\value{
data.frame with UIDs for the synonyms on GenBank
}
\description{
Searches first for the synonyms and then searches for the UID code of the synonym present on GenBank
}
\details{
The function is a wraper function of syno_mycobank and syns_on_ncbi.
}
\examples{
taxon <- "Friesia annosa"
search_syn_in_ncbi(taxon = taxon)
}
\author{
Franz-Sebastian Krah
}

