% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syno_mycobank.R
\name{syno_mycobank}
\alias{syno_mycobank}
\title{Downloads synonyms from mycobank.org}
\usage{
syno_mycobank(taxon)
}
\arguments{
\item{taxon}{a \code{character} containing fungal species name}
}
\value{
vector of class \code{character}.
}
\description{
Searches and downloads synonmys from the mycobank.org website for a given fungal species name
}
\details{
The function searches all synonym entries on the MycoBank on-line database. The data may then be further processes, e.g. a search which of the synonmys are present at GenBank (syns_on_ncbi).
}
\examples{
syns <- syno.mycobank("Heterobasidion annosum")
}
\author{
Franz-Sebastian Krah
}

