% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syns_on_ncbi.R
\name{syns_on_ncbi}
\alias{syns_on_ncbi}
\title{Checks synonyms against GenBank}
\usage{
syns_on_ncbi(synonyms)
}
\arguments{
\item{synonyms}{a vector of class \code{character} containing fungal species names}
}
\value{
an data.frame with the synomyms and the GIs, if present.
}
\description{
Searches which of the synonyms of an input species list are present on GenBank
}
\details{
The function searches which of the synonmys are present at GenBank. If you have a species list e.g. from a field study and want to derive sequences from GenBank but only get a subset of the species; then you might ignore the synonyms.
}
\examples{
synonyms <- c("Heterobasidion annosum", "Polyporus annosus","Polyporus subpileatus","Polyporus scoticus","Polyporus makraulos","Polyporus macraulos","Trametes radiciperda","Poria macraula","Poria macraula","Polyporus irregularis","Polystictoides fuscus","Polyporus atramosus","Polyporus marginatoides","Polyporus atrannosus","Heterobasidion annosum f. macraulos")
gi <- syns_on_ncbi(synonyms)
}
\author{
Franz-Sebastian Krah
}

